sonarr-info/requirements.txt
arrapi
plexapi
ruamel.yaml